INT-46: ipam.Address.type is now ipam.Address.status
The "type" field of ipam.Address has been renamed to "status" in
the API as of v2.2. Additionally, the "assignment" value becomes
"assigned", and the "host" value is dropped.

For those using IPAM functionality, there are some minor but
potentially breaking changes:

v2.1 users:
  * please continue to use v0.13.0 of the SDK

v2.2 users with existing code for IPAM:
  * NS1 class:
    * createAddress method: the "type" arg is now "status"
    * loadAddressByPrefix method: the "type" arg is now "status"
  * ipam.Address:
    * "type" kwarg to init is now "status"
    * "type" attr is now "status" attr
  * ipam.Network:
    * new_address method: "type" arg is now "status"
  * Valid values for "status" (née "type") have changed:
    * previously: ["planned", "assignment", "host"]
    * now:        ["planned", "assigned"]