INT-129: support for "rate-limit strategies"
The existing rate limit handling is based on handling an exception when
we get a 429 response. We add some strategies that base behavior on the headers
from every response, without waiting for a 429. These are the same strategies
we have in the Go-SDK:

Adds two strategies:
  * solo - sleep longer the closer we are to running out of tokens
  * concurrent - needs a "parallelism" argument, accounts for the other workers
    in determining how long to sleep. doesn't kick in until tokens remaining
    is <= parallelism

Strategy is set and determined by config variables, the default (for now) is
"no strategy", i.e. preserves current behavior.

Added tests for resource: exercises transports and rate-limiting

There is an example of using rate-limiting strategies

I added support for these via config, and the different "transports". There is
some refactoring of transport internals to fix bugs, support various py
versions, and for ease of testing. Functionality should not be affected
in normal use:

* basic transport
  * pull ssl workaround for url openers into a method
  * replace references to `resp.hdrs` with a _get_headers function that
    should be consistent across py versions.
  * try/except around body.decode to handle str/bytes response from different
    py versions
* twisted transport
  * move ResourceException into class init. having it at top-level makes it
    tough to use e.g. py3.3 (without twisted).

Bonus:

* Update README about supported python versions and what we run CI against.
  turns out the "py33" env in tox.ini is actually 3.6